+++ guidata/__init__.py
@@ -8,7 +8,7 @@ user interfaces for easy dataset editing and display. It also provides helpers
 and application development tools for Qt.
 """
 
-__version__ = "3.5.3"
+__version__ = "3.5.4"
 
 
 # Dear (Debian, RPM, ...) package makers, please feel free to customize the
@@ -30,7 +30,9 @@ def qapplication():
     if not app:
         app = QApplication([])
         install_translator(app)
-        set_color_mode(app)
+        from guidata import qthelpers
+
+        qthelpers.set_color_mode()
     return app
 
 
@@ -53,39 +55,3 @@ def install_translator(qapp):
                 break
     if QT_TRANSLATOR is not None:
         qapp.installTranslator(QT_TRANSLATOR)
-
-
-def set_color_mode(app):
-    """Set color mode (dark or light), depending on OS setting"""
-    from qtpy.QtCore import Qt
-    from qtpy.QtGui import QColor, QPalette
-    from qtpy.QtWidgets import QStyleFactory
-
-    from guidata import qthelpers
-
-    if qthelpers.is_dark_mode():
-        app.setStyle(QStyleFactory.create("Fusion"))
-        dark_palette = QPalette()
-        dark_color = QColor(50, 50, 50)
-        disabled_color = QColor(127, 127, 127)
-        dpsc = dark_palette.setColor
-        dpsc(QPalette.Window, dark_color)
-        dpsc(QPalette.WindowText, Qt.white)
-        dpsc(QPalette.Base, QColor(31, 31, 31))
-        dpsc(QPalette.AlternateBase, dark_color)
-        dpsc(QPalette.ToolTipBase, Qt.white)
-        dpsc(QPalette.ToolTipText, Qt.white)
-        dpsc(QPalette.Text, Qt.white)
-        dpsc(QPalette.Disabled, QPalette.Text, disabled_color)
-        dpsc(QPalette.Button, dark_color)
-        dpsc(QPalette.ButtonText, Qt.white)
-        dpsc(QPalette.Disabled, QPalette.ButtonText, disabled_color)
-        dpsc(QPalette.BrightText, Qt.red)
-        dpsc(QPalette.Link, QColor(42, 130, 218))
-        dpsc(QPalette.Highlight, QColor(42, 130, 218))
-        dpsc(QPalette.HighlightedText, Qt.black)
-        dpsc(QPalette.Disabled, QPalette.HighlightedText, disabled_color)
-        app.setPalette(dark_palette)
-        app.setStyleSheet(
-            "QToolTip { color: #ffffff; background-color: #2a82da; border: 1px solid white; }"
-        )